+++ src/main/java/soot/dexpler/instructions/ConstInstruction.java
@@ -41,6 +41,9 @@ import soot.jimple.Constant;
 import soot.jimple.IntConstant;
 import soot.jimple.Jimple;
 import soot.jimple.LongConstant;
+import soot.Local;
+import soot.dexpler.DexTypeInference;
+import soot.UnknownType;
 
 public class ConstInstruction extends DexlibAbstractInstruction {
 
@@ -53,7 +56,18 @@ public class ConstInstruction extends DexlibAbstractInstruction {
         int dest = ((OneRegisterInstruction) instruction).getRegisterA();
 
         Constant cst = getConstant(dest, body);
-        AssignStmt assign = Jimple.v().newAssignStmt(body.getRegisterLocal(dest), cst);
+        Local target;
+        // NOTE hzh<huzhenghao@sbrella.com>: Dex doesnt differenciate Int typed 0 and Pointer
+        // Type. Propogate the register type explcitly to Unknown, so it will get a second
+        // chance to be inferred to a correct type.
+        if (cst.equals(IntConstant.v(0))
+           // This is to avoid reset the type of reg after it has initialized - One case is
+           // reg type initialization based on local variable Debug Info
+           && body.getRegisterLocal(dest).getType() instanceof UnknownType)
+            target = DexTypeInference.applyForward(dest, UnknownType.v(), body);
+        else
+            target = DexTypeInference.applyForward(dest, cst.getType(), body);
+        AssignStmt assign = Jimple.v().newAssignStmt(target, cst);
         setUnit(assign);
         addTags(assign);
         body.add(assign);