+++ src/main/java/soot/dexpler/instructions/MoveExceptionInstruction.java
@@ -36,6 +36,8 @@ import soot.dexpler.IDalvikTyper;
 import soot.dexpler.typing.DalvikTyper;
 import soot.jimple.IdentityStmt;
 import soot.jimple.Jimple;
+import soot.jimple.CaughtExceptionRef;
+import soot.dexpler.DexTypeInference;
 
 public class MoveExceptionInstruction extends DexlibAbstractInstruction implements RetypeableInstruction {
 
@@ -49,8 +51,9 @@ public class MoveExceptionInstruction extends DexlibAbstractInstruction implemen
     @Override
 	public void jimplify (DexBody body) {
         int dest = ((OneRegisterInstruction)instruction).getRegisterA();
-        Local l = body.getRegisterLocal(dest);
-        stmtToRetype = Jimple.v().newIdentityStmt(l, Jimple.v().newCaughtExceptionRef());
+        CaughtExceptionRef caughtRef = Jimple.v().newCaughtExceptionRef();
+        Local l = DexTypeInference.applyForward(dest, caughtRef.getType(), body);
+        stmtToRetype = Jimple.v().newIdentityStmt(l, caughtRef);
         setUnit(stmtToRetype);
         addTags(stmtToRetype);
         body.add(stmtToRetype);